requirements.txt
PyQt5==5.15.10
openai-whisper>=20231117
faster-whisper>=1.0.3
sounddevice>=0.4.7
pynput>=1.7.7
pyautogui>=0.9.54
numpy>=1.24.0,<2.0
psutil>=5.9.8
pywin32>=306
torch>=2.0.0,<2.2.0
